working, kind of

+++ robot.py
@@ -1,3 +1,66 @@
+# import os 
+# import math
+# import pybullet as p
+# import pyrosim.pyrosim as pyrosim
+# from pyrosim.neuralNetwork import NEURAL_NETWORK
+
+# from sensor import SENSOR
+# from motor import MOTOR
+
+# import constants as c 
+
+
+
+
+# class ROBOT:
+#     def __init__(self,solutionID, links, joints, delete=True) -> None:
+#         self.solutionID = solutionID
+#         self.sensors = {}
+#         self.motors = {}
+#         self.robotId = p.loadURDF("body.urdf")
+#         self.nn = NEURAL_NETWORK("brain{}.nndf".format(solutionID))
+#         pyrosim.Prepare_To_Simulate(self.robotId)
+#         self.Prepare_To_Sense()
+#         self.Prepare_to_Act()
+#         os.system("rm brain{}.nndf".format(solutionID))
+        
+
+#     def Prepare_To_Sense(self):
+#         for linkName in pyrosim.linkNamesToIndices:
+#             self.sensors[linkName] = SENSOR(linkName)
+    
+#     def Sense(self,t):
+#         for sensor in self.sensors:
+#             self.sensors[sensor].Get_Value(t)
+
+#     def Prepare_to_Act(self):
+#         for jointName in pyrosim.jointNamesToIndices:
+#             self.motors[jointName] = MOTOR(jointName)
+
+#     def Act(self,t):
+#         for neuronName in self.nn.Get_Neuron_Names():
+#             if self.nn.Is_Motor_Neuron(neuronName):
+#                 jointName = self.nn.Get_Motor_Neurons_Joint(neuronName)
+#                 desiredAngle = self.nn.Get_Value_Of(neuronName) * c.motorJointRange
+#                 self.motors[jointName].Set_Value(self.robotId,desiredAngle)
+
+
+#     def Think(self):
+#         self.nn.Update()
+
+#     def Get_Fitness(self):
+#         basePositionAndOrientation = p.getBasePositionAndOrientation(self.robotId)
+#         basePosition = basePositionAndOrientation[0]
+#         xCoordinateOfLinkZero = basePosition[0]
+#         yCoord = basePosition[2]
+#         zCoord = basePosition[1]
+
+#         dist_squared = ((xCoordinateOfLinkZero - (-10))**2 + (zCoord - (10))**2 + (yCoord - (2))**2)
+#         dist = math.sqrt(dist_squared)
+#         with open("tmp{}.txt".format(self.solutionID), 'w') as f:
+#             f.write(str(dist))
+#         os.system("mv tmp{}.txt fitness{}.txt".format(self.solutionID,self.solutionID))
+
 import os 
 import math
 import pybullet as p
@@ -13,50 +76,79 @@ import constants as c
 
 
 class ROBOT:
-    def __init__(self,solutionID) -> None:
+    def __init__(self,solutionID,links, joints, delete=True) -> None:
         self.solutionID = solutionID
+        self.links = links
+        self.joints = joints
         self.sensors = {}
         self.motors = {}
-        self.robotId = p.loadURDF("body.urdf")
+        try:
+            self.robotId = p.loadURDF("body{}.urdf".format(solutionID))
+        except:
+            self.write_0_fitness()
+            return None
         self.nn = NEURAL_NETWORK("brain{}.nndf".format(solutionID))
         pyrosim.Prepare_To_Simulate(self.robotId)
         self.Prepare_To_Sense()
         self.Prepare_to_Act()
-        os.system("rm brain{}.nndf".format(solutionID))
+        if delete:
+            pass
+            os.system("rm brain{}.nndf".format(solutionID))
+            os.system("rm body{}.urdf".format(solutionID))
         
 
     def Prepare_To_Sense(self):
-        for linkName in pyrosim.linkNamesToIndices:
-            self.sensors[linkName] = SENSOR(linkName)
+        for linkName in self.links:
+                self.sensors[linkName] = SENSOR(linkName)
     
     def Sense(self,t):
         for sensor in self.sensors:
-            self.sensors[sensor].Get_Value(t)
+            try:
+                self.sensors[sensor].Get_Value(t)
+            except:
+                pass
 
     def Prepare_to_Act(self):
-        for jointName in pyrosim.jointNamesToIndices:
-            self.motors[jointName] = MOTOR(jointName)
+        for jointName in self.joints:
+            try:
+                self.motors[jointName] = MOTOR(jointName)
+            except:
+                pass
 
     def Act(self,t):
-        for neuronName in self.nn.Get_Neuron_Names():
-            if self.nn.Is_Motor_Neuron(neuronName):
-                jointName = self.nn.Get_Motor_Neurons_Joint(neuronName)
-                desiredAngle = self.nn.Get_Value_Of(neuronName) * c.motorJointRange
-                self.motors[jointName].Set_Value(self.robotId,desiredAngle)
+        try:
+            for neuronName in self.nn.Get_Neuron_Names():
+                if self.nn.Is_Motor_Neuron(neuronName):
+                    jointName = self.nn.Get_Motor_Neurons_Joint(neuronName)
+                    desiredAngle = self.nn.Get_Value_Of(neuronName) * c.motorJointRange
+                    self.motors[jointName].Set_Value(self.robotId,desiredAngle)
+        except:
+            pass
 
 
     def Think(self):
-        self.nn.Update()
+        try:
+            self.nn.Update()
+        except:
+            pass
 
     def Get_Fitness(self):
-        basePositionAndOrientation = p.getBasePositionAndOrientation(self.robotId)
-        basePosition = basePositionAndOrientation[0]
-        xCoordinateOfLinkZero = basePosition[0]
-        yCoord = basePosition[2]
-        zCoord = basePosition[1]
+        try:
+            basePositionAndOrientation = p.getBasePositionAndOrientation(self.robotId)
+            basePosition = basePositionAndOrientation[0]
+            xCoordinateOfLinkZero = basePosition[0]
+            yCoord = basePosition[2]
+            zCoord = basePosition[1]
 
-        dist_squared = ((xCoordinateOfLinkZero - (-10))**2 + (zCoord - (10))**2 + (yCoord - (2))**2)
-        dist = math.sqrt(dist_squared)
+            dist_squared = ((xCoordinateOfLinkZero - (0))**2 + (zCoord - (0))**2) # + (yCoord - (0))**2)
+            dist = math.sqrt(dist_squared)
+        except:
+            dist = 0 
         with open("tmp{}.txt".format(self.solutionID), 'w') as f:
             f.write(str(dist))
+        os.system("mv tmp{}.txt fitness{}.txt".format(self.solutionID,self.solutionID))
+    
+    def write_0_fitness(self):
+        with open("tmp{}.txt".format(self.solutionID), 'w') as f:
+            f.write(str('0.00'))
         os.system("mv tmp{}.txt fitness{}.txt".format(self.solutionID,self.solutionID))
